Moved music related events to Music cog and updated code for wavelink 3.2.0.
